Modify curve and surface classes to use evaluators
An Evaluator consists of a single or multiple interchangable algorithms
for computing curve and surface points, and in future, derivatives.

Evaluators are designed to change the evaluation subsystem (i.e. the
algorithm that helps the library to solve the problem) of the class and
they can be replaced with each other at runtime (Strategy pattern).