API stabilization updates
Unfortunately, basing the API on the shapes gives some problems in
common end points (i.e. properties). In order to get a stable API, the
return value data exporting end point must be standardized, here "data"
property is modified to return a dict of lists on degree and  knot
vector values regardless of the shape. Also, the storage types of the
private fields are updated